requirements.txt
Django==5.2.4
crispy-forms==2.1
crispy-bootstrap5==0.7
Pillow==10.4.0
qrcode==7.4.2
django-ratelimit==4.1.0
django-crontab==0.7.1
python-dotenv==1.0.0
dj-database-url==2.1.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
redis==5.0.1
aiohttp==3.9.1
PyMuPDF==1.23.14
pdf2image==1.17.0 